RW-15: add loader when starting camera
