+++ upload-app/src/common/components/index.ts
@@ -4,6 +4,7 @@ export * from './Button';
 export * from './Header';
 export * from './Input';
 export * from './Loader';
+export * from './LoaderBox';
 export * from './Section';
 export * from './Summary';
 export * from './SummaryItem';
